ml_service/preprocess/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnslib==0.9.14
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
pkg-resources==0.0.0
py4j==0.10.9
pymongo==3.11.0
pyspark==3.0.0
python-dateutil==2.8.1
PythonDNS==0.1
pytz==2020.1
requests==2.24.0
requests-file==1.5.1
six==1.15.0
soupsieve==2.0.1
tldextract==2.2.3
urllib3==1.25.10
Werkzeug==1.0.1
whois==0.9.7
